+++ src/Screens/MusicSelect/MusicSelect.hpp
@@ -47,7 +47,7 @@ namespace MusicSelect {
         // converts a mouse click into a button press
         void handle_mouse_click(const sf::Event::MouseButtonEvent& mouse_button_event);
         // chooses what happens for each button
-        void press_button(const Input::Button& button);
+        void press_button(Input::Button button);
 
         void cycle_display_styles(sf::RenderWindow& window);