+++ src/Screens/Results/Resources.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "../../Resources/SharedResources.hpp"
+
+namespace Results {
+    struct ScreenResources : Resources::HoldsSharedResources {
+        ScreenResources(Resources::SharedResources& r) : Resources::HoldsSharedResources(r) {};
+    };
+
+    struct HoldsResources : Resources::HoldsSharedResources {
+        HoldsResources(ScreenResources& t_resources) : Resources::HoldsSharedResources(t_resources.shared), resources(t_resources) {};
+        ScreenResources& resources;
+    };
+}